Support direct connections to RabbitMQ
This opens path to using this library for writing RabbitMQ plugins.
Or as a performance improvements, escpecially when RabbitMQ is being
run in embedded mode.

Sadly there is no way to provide a reliable tests for this feature, as
we need to know both the Erlang node-name of the broker and the Erlang
cookie. And e.g. for Travis with disabled `sudo` it's not possible.